+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `MollieEx.Permissions.list/2` and `MollieEx.Permissions.get/3`.
 - Add `MollieEx.Organizations.get/3`, `MollieEx.Organizations.current/2`,
   and `MollieEx.Organizations.partner_status/2`.
+- Add `MollieEx.Onboarding.get/2`.
 
 ## 0.4.0 - 2026-06-14